Split the video stage, and make its feed figures per picture
`stage.rs` was 770 lines holding three separable things. Now
`stage/{mod,metrics,parts}.rs`: the latency arithmetic with its tests, the
slice-progressive reassembler, and `VideoStage` itself. Same idiom as
`app/render/` and `ndl/`. No behaviour change in the move.

Two figures were wrong on a slice-progressive session, where `feed` runs once per
PIECE rather than once per AU:

- `stats.feed_us` stored the last slice's submit time instead of the AU's. Now
  accumulated across the pieces, and dropped when an AU is abandoned so it can't
  leak into the next one.
- Nothing said whether parts were being fed at all. The video heartbeat now logs
  `parts=` beside `frames=`; core emits an early part only for an AU spanning
  more than one FEC block, so on a mode with small AUs the whole lever is inert
  and that ratio is the only way to see it.

Also fixes `connect`'s pointer to `AuParts`, which has not been in `session::pump`
for some time.
